rename server.js in index.js
